Removed ip_location example as it has a bunch of strange "Special" operators hardcoded into the Hadoop library just for it's own usage. Should be readded using proper MongoDB syntax.
At this time, we don't support in place updates of Data with the Hadoop
driver and hardcoding "special" operators serves only to confuse users.

There is an open TODO item to investigate how to support incremental
map/reduce, but updating values in place is not an appropriate way to
handle a MapReduce approach.