Add capability of BSONLoader.java to parse UUID
If the type of the object passed to convertBSONtoPigType UUID it is not recognised; the method defaults to outputting the object and breaking pig. If the type is objectid where the semantics are similar, the method already calls toString() on the object. This change just adds similar ability to the UUID object.

For further information see:

http://stackoverflow.com/questions/17579027/pig-mongoloader-exception-loading-data-with-uuids/21391717#21391717