Add support for PySpark (HADOOP-187).
This adds a "spark" module to the project, which compiles into
"mongo-hadoop-spark.jar".  This jar is currently only necessary if you want to
use PySpark with mongo-hadoop.  Additionally, this adds the pymongo_spark
module, which provides the necessary objects and methods to use mongo-hadoop and
PySpark together on the Python-side.